Bugfix for dds_read() undefined behaviour
ubsan reported an error in dds_read() on my HPS because it was trying to left-shift the value 163 by 24 places. This is not possible to do with a 32-bit signed integer.
The code had several static analyzer warnings related to the use of signed integers as well.
Created a new function to do the conversion, which uses unsigned types exclusively.